# backend/ledgerguard-api/src/main/java/com/ledgerguard/reconciliation/api/ReconciliationDashboardSummaryResponse.java
package com.ledgerguard.reconciliation.api;

public record ReconciliationDashboardSummaryResponse(
        long openCases,
        long inReviewCases,
        long resolvedCases,
        long totalCases,
        long totalRuns,
        ReconciliationRunSummaryResponse latestRun
) {
}