# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LedgerGuard Payment Integrity & Ledger Platform API",
    "description": "High-reliability financial core implementing immutable double-entry ledgering, deterministic concurrency control, transactional outbox messaging, multi-level reconciliation, and ambiguous external payment recovery.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/reconciliation/cases/{caseId}/resolve": {
      "post": {
        "tags": [
          "reconciliation-controller"
        ],
        "description": "**Required Roles**: `ROLE_OPS` (Bearer JWT). Customers and Merchants are strictly forbidden.",
        "operationId": "resolveCase",
        "parameters": [
          {
            "name": "caseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconciliationCaseResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-required-roles": [
          "OPS"
        ]
      }
    },
    "/api/reconciliation/cases/{caseId}/repair-snapshot": {
      "post": {
        "tags": [
          "reconciliation-controller"
        ],
        "description": "**Required Roles**: `ROLE_OPS` (Bearer JWT). Customers and Merchants are strictly forbidden.",
        "operationId": "repairSnapshot",
        "parameters": [
          {
            "name": "caseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotRepairResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-required-roles": [
          "OPS"
        ]
      }
    },
    "/api/reconciliation/cases/{caseId}/claim": {
      "post": {
        "tags": [
          "reconciliation-controller"
        ],
        "description": "**Required Roles**: `ROLE_OPS` (Bearer JWT). Customers and Merchants are strictly forbidden.",
        "operationId": "claimCase",
        "parameters": [
          {
            "name": "caseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconciliationCaseResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-required-roles": [
          "OPS"
        ]
      }
    },
...
    "/api/reconciliation/runs": {
      "get": {
        "tags": [
          "reconciliation-controller"
        ],
        "description": "**Required Roles**: `ROLE_OPS` (Bearer JWT). Customers and Merchants are strictly forbidden.",
        "operationId": "getRuns",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseReconciliationRunSummaryResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-required-roles": [
          "OPS"
        ]
      },
      "post": {
        "tags": [
          "reconciliation-controller"
        ],
        "summary": "Trigger on-demand synchronous reconciliation run",
        "description": "**Required Roles**: `ROLE_OPS` (Bearer JWT). Customers and Merchants are strictly forbidden.",
        "operationId": "triggerRun",
        "responses": {
          "201": {
            "description": "Created",
            "headers": {
              "Location": {
                "description": "Relative URI pointing to the newly created reconciliation run resource",
                "schema": {
                  "type": "string",
                  "example": "/api/reconciliation/runs/550e8400-e29b-41d4-a716-446655440000"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconciliationRunSummaryResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Reconciliation is already running",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-required-roles": [
          "OPS"
        ]
      }
    },
    "/api/reconciliation/runs/{runId}": {
      "get": {
        "tags": [
          "reconciliation-controller"
        ],
        "description": "**Required Roles**: `ROLE_OPS` (Bearer JWT). Customers and Merchants are strictly forbidden.",
        "operationId": "getRunById",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconciliationRunSummaryResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-required-roles": [
          "OPS"
        ]
      }
    },
    "/api/reconciliation/runs/{runId}/items": {
      "get": {
        "tags": [
          "reconciliation-controller"
        ],
        "description": "**Required Roles**: `ROLE_OPS` (Bearer JWT). Customers and Merchants are strictly forbidden.",
        "operationId": "getRunItems",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseReconciliationItemResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-required-roles": [
          "OPS"
        ]
      }
    },
    "/api/reconciliation/cases": {
      "get": {
        "tags": [
          "reconciliation-controller"
        ],
        "description": "**Required Roles**: `ROLE_OPS` (Bearer JWT). Customers and Merchants are strictly forbidden.",
        "operationId": "getCases",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "OPEN",
                "IN_REVIEW",
                "RESOLVED"
              ]
            }
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "JOURNAL_BALANCE",
                "SNAPSHOT_CONSISTENCY",
                "PROVIDER_SETTLEMENT"
              ]
            }
          },
          {
            "name": "classification",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "DISCREPANCY",
                "UNRESOLVED"
              ]
            }
          },
          {
            "name": "problemType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "UNBALANCED_JOURNAL",
                "MALFORMED_JOURNAL",
                "SNAPSHOT_MISMATCH",
                "SNAPSHOT_MISSING",
                "PROVIDER_STATUS_MISMATCH",
                "PROVIDER_IDENTITY_MISMATCH",
                "PROVIDER_NOT_FOUND",
                "PROVIDER_UNAVAILABLE",
                "PROVIDER_STILL_PROCESSING"
              ]
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "runId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "assigned",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseReconciliationCaseResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-required-roles": [
          "OPS"
        ]
      }
    },
    "/api/reconciliation/cases/{caseId}": {
      "get": {
        "tags": [
          "reconciliation-controller"
        ],
        "description": "**Required Roles**: `ROLE_OPS` (Bearer JWT). Customers and Merchants are strictly forbidden.",
        "operationId": "getCaseById",
        "parameters": [
          {
            "name": "caseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconciliationCaseResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-required-roles": [
          "OPS"
        ]
      }
    },
...
    "/api/payments/summary": {
      "get": {
        "tags": [
          "payment-controller"
        ],
        "description": "**Required Roles**: `ROLE_MERCHANT` (Bearer JWT). `CUSTOMER` and `OPS` are forbidden.",
        "operationId": "summary",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantSummaryResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-required-roles": [
          "MERCHANT"
        ]
      }
    },
...
    "/api/reconciliation/summary": {
      "get": {
        "tags": [
          "reconciliation-controller"
        ],
        "summary": "Get authoritative reconciliation dashboard summary metrics",
        "description": "**Required Roles**: `ROLE_OPS` (Bearer JWT). Customers and Merchants are strictly forbidden.",
        "operationId": "getSummary",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconciliationDashboardSummaryResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-required-roles": [
          "OPS"
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ManualResolveRequest": {
        "type": "object",
        "properties": {
          "resolutionNote": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 0
          }
        },
        "required": [
          "resolutionNote"
        ]
      },
      "ReconciliationCaseResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "reconciliationItemId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "assignedToUserId": {
            "type": "string",
            "format": "uuid"
          },
          "resolvedByUserId": {
            "type": "string",
            "format": "uuid"
          },
          "resolutionAction": {
            "type": "string"
          },
          "resolutionNote": {
            "type": "string"
          },
          "openedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "resolvedAt": {
            "type": "string",
            "format": "date-time"
          },
          "item": {
            "$ref": "#/components/schemas/ReconciliationItemResponse"
          }
        }
      },
      "ReconciliationItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "reconciliationRunId": {
            "type": "string",
            "format": "uuid"
          },
          "classification": {
            "type": "string"
          },
          "level": {
            "type": "string"
          },
          "problemType": {
            "type": "string"
          },
          "entityType": {
            "type": "string"
          },
          "entityId": {
            "type": "string",
            "format": "uuid"
          },
          "observedLocalStatus": {
            "type": "string"
          },
          "expectedValue": {
            "type": "string"
          },
          "actualValue": {
            "type": "string"
          },
          "providerStatus": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "detectedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SnapshotRepairResponse": {
        "type": "object",
        "properties": {
          "caseId": {
            "type": "string",
            "format": "uuid"
          },
          "ledgerAccountId": {
            "type": "string",
            "format": "uuid"
          },
          "previousBalanceMinor": {
            "type": "string"
          },
          "repairedBalanceMinor": {
            "type": "string"
          },
          "resolutionAction": {
            "type": "string"
          },
          "snapshotUpdatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PagedResponseReconciliationRunSummaryResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReconciliationRunSummaryResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ReconciliationRunSummaryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "triggerSource": {
            "type": "string"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time"
          },
          "journalsChecked": {
            "type": "integer",
            "format": "int64"
          },
          "accountsChecked": {
            "type": "integer",
            "format": "int64"
          },
          "operationsChecked": {
            "type": "integer",
            "format": "int64"
          },
          "discrepancyCount": {
            "type": "integer",
            "format": "int64"
          },
          "unresolvedCount": {
            "type": "integer",
            "format": "int64"
          },
          "failureReason": {
            "type": "string"
          }
        }
      },
      "PagedResponseReconciliationItemResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReconciliationItemResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PagedResponseReconciliationCaseResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReconciliationCaseResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "MerchantSummaryResponse": {
        "type": "object",
        "properties": {
          "grossReceivedMinor": {
            "type": "string"
          },
          "platformFeesMinor": {
            "type": "string"
          },
          "netReceivedMinor": {
            "type": "string"
          },
          "refundedAmountMinor": {
            "type": "string"
          },
          "pendingPayoutsCount": {
            "type": "integer",
            "format": "int64"
          },
          "pendingPayoutsAmountMinor": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          }
        }
      },
...
      "ReconciliationDashboardSummaryResponse": {
        "type": "object",
        "properties": {
          "openCases": {
            "type": "integer",
            "format": "int64"
          },
          "inReviewCases": {
            "type": "integer",
            "format": "int64"
          },
          "resolvedCases": {
            "type": "integer",
            "format": "int64"
          },
          "totalCases": {
            "type": "integer",
            "format": "int64"
          },
          "totalRuns": {
            "type": "integer",
            "format": "int64"
          },
          "latestRun": {
            "$ref": "#/components/schemas/ReconciliationRunSummaryResponse"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Stateless JWT access token. Required for authenticated user, customer, merchant, and ops endpoints.",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "refreshCookie": {
        "type": "apiKey",
        "description": "Opaque refresh token delivered via HttpOnly, SameSite=Strict cookie for session rotation and logout.",
        "name": "ledgerguard_refresh_token",
        "in": "cookie"
      },
      "pspWebhookSignature": {
        "type": "apiKey",
        "description": "HMAC-SHA256 signature payload verification header for external PSP webhook events.",
        "name": "X-PSP-Webhook-Signature",
        "in": "header"
      },
      "pspWebhookTimestamp": {
        "type": "apiKey",
        "description": "Unix epoch millisecond timestamp header for replay attack protection.",
        "name": "X-PSP-Webhook-Timestamp",
        "in": "header"
      }
    }
  }
}